Explicitly set serverSelectionTimeoutMS for 'fallback' Mongo connection string
When the tests are executed a number of attempts are made to initialise the
MongoClient before the Mongo instance has been initialised. By default, this
triggers a 30000ms timeout for each attempt.

The 'fallback' Mongo connection string is only used when the
Mongo user & password environment variables are not passed and therefore
only applies for local development/testing use-cases.

This change shortens the timeout to 100ms reducing overall test suite execution
time from ~40 mins to ~8 mins.